Updated Packager to better version that uses online Google Closure Compiler.
