+++ minify.php
@@ -1,5 +1,5 @@
 <?php
-require_once 'bin/Packager/Packager.php';
+require_once 'bin/Packager.php';
 
 $jsSuccess = Packager::js(array(
 	'src/jcsdl.loader.js',
